Refactored objects.replib, related cython & native python modules into subpackage
