+++ .gitignore
@@ -50,8 +50,8 @@ packages/pyGSTi.egg-info
 parsetab_string.py
 
 # These are handwritten, not generated by cython, and should not be ignored
-!packages/pygsti/objects/fastreps.cpp
-!packages/pygsti/objects/fastreps.h
+!packages/pygsti/objects/replib/fastreps.cpp
+!packages/pygsti/objects/replib/fastreps.h
 
 # Packages #
 ############
