+++ packages/pygsti/objects/replib/fastreps.cpp
